requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask-cors==5.0.1
stripe==12.1.0
gunicorn==21.2.0 